+++ xxc/app/core/commander.js
@@ -120,6 +120,13 @@ export const executeCommand = (command, ...params) => {
  * @return {Promise<any, Error>} 通过 Promise 返回命令执行结果 (Return result with Promise)
  */
 export const executeCommandLine = (commandLine, commandContext = null) => {
+    if (commandLine.includes('!')) {
+        return commandLine.split('!').forEach(cLine => {
+            if (cLine) {
+                executeCommandLine(cLine, commandContext);
+            }
+        });
+    }
     setCommandContext(commandContext);
     const params = commandLine.split('/');
     return executeCommand(...params.map((p, idx) => {